pyproject.toml
[tool.poetry]
name = "time-wise-rat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.8"
torch = "^2.2.1"
torchvision = "^0.17.1"
torchaudio = "^2.2.1"
pandas = "^2.2.1"
jupyter = "^1.0.0"
matplotlib = "^3.8.3"
safetensors = "^0.4.2"
tqdm = "^4.66.2"
lightning = "^2.2.1"
scikit-learn = "^1.4.1.post1"
tensorboard = "^2.16.2"
faiss-cpu = "^1.8.0"
hydra-core = "^1.3.2"
dtw-python = "^1.4.2"
pynndescent = "^0.5.12"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
